hardening script / chrome installation

+++ ods-devenv/scripts/deploy.sh
@@ -118,9 +118,12 @@ EOF
 #   None
 #######################################
 function setup_google_chrome() {
-    curl -LO https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
-    sudo yum install -y ./google-chrome-stable_current_*.rpm
-    rm ./google-chrome-stable_current_*.rpm
+    if [[ -z $(command -v google-chrome) ]]
+    then
+        curl -LO https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
+        sudo yum install -y ./google-chrome-stable_current_*.rpm
+        rm ./google-chrome-stable_current_*.rpm
+    fi
 }
 
 #######################################
